+++ src/main/java/edu/jhuapl/trinity/utils/umap/Utils.java
@@ -103,7 +103,7 @@ final class Utils {
         for (final double v : vec) {
             result += v * v;
         }
-        return (double) Math.sqrt(result);
+        return Math.sqrt(result);
     }
 
     /**
